requirements.txt
# Librerías necesarias para el proyecto
numpy
matplotlib
opencv-python
python-dotenv